add a <vaadin-router-demo-snipptet> element to demo the URL changes
The new `<vaadin-router-demo-snippet>` element is based off of the `<vaadin-demo-snippet>` element from the vaadin-demo-helpers repo. The difference is that instead of inserting the demo code into its own shadow DOM, the new element inserts the demo code into an iframe. That allows each demo snippet have its own window.location.

There is a message-based communication channel between the iframe and the demo snippet element to allow the demo snippet to show and manipulate the iframe's URL.