allow dynamic changes to the route template
When the route content is defined by an inline template, this inline template might be added with a delay (due to a slow network) or dynamically (from JS). This commit uses a mutation observer to detect changes to the inline template if they happen after the initial render. Template changes are treated as any other change in the route parameters (i.e. the current location and the route URL).