add support for route paramters (e.g. /users/:user)  (#2)
* allow dynamic changes to the route template

When the route content is defined by an inline template, this inline template might be added with a delay (due to a slow network) or dynamically (from JS). This commit uses a mutation observer to detect changes to the inline template if they happen after the initial render. Template changes are treated as any other change in the route parameters (i.e. the current location and the route URL).

* add support for route paramters (e.g. /users/:user)

The implementation is based on the MIT-licensed path-to-regexp library.
Features:
 - named parameters (e.g. /users/:user/messages/:messageId)
 - optional parameters marked with a `?` (e.g. /images/:size/:color? matches both /images/100x100 and /images/100x100/hotpink)
 - multi-segment parameters (zero or more marked with a `*`, one or more marked with a `+`) (e.g. /kb/:path+ matches /kb/general/insurance/personal and /kb/general)
 - regex patterns for parameter values (e.g. /icons/:width(\\d+)x:height(\\d+) matches /icons/96x96)
 - unnamed parameters (e.g. /users/(.*) matches /users/asdf)

* add a demo section for route parameters

* add a demo for optional route paramters

* update the styles for rendered demo snippets

* add a demo for multi-segment route parameters

* rename vaadin-route url property to path

- to decouple the route path from a URL because the path could have placeholders (which is not a valid URL)