set up Travis CI: lint, test, coverage, and bundle size (#60)
Both in the router and the router-core packages:
 - `yarn lint` runs a linter and fails the build on errors. The linter configs are `.eslintrc`, `.stylelintrc` and the default `polymer-2` ruleset of the polymer CLI linter.
 - `yarn test:coverage` runs unit tests in a local chrome, reports the test coverage (with istanbul), and fails the build if the coverage is below the threshold set in `wct.conf.js`
 - `yarn test` runs integration tests in all local browsers (via WCT) and fails the build on failure
 - `yarn size-control` calculates the effective library size (minified, gzipped) and fails the build if it exeeds the threshold set in `package.json`.

All 4 NPM scripts are executed on Travis for both the core and the element modules. The coverage thresholds are set to 80% for the core and 50% for the element (to make the builds pass, since the work on the element is not started and there is not enough tests there). The bundle size limits are set to 5kB and 10kB minified gzipped for the core and the element modules respectively.

Fixes #12