limit the Travis 'build on push' feature to the master branch only
Without this setting each time a new commit is pushed to a PR branch Travis triggers two identical builds:
 - build on a PR
 - build on a push

This setting chanes the behavior so that only one build is triggered. The downside is that pushes to non-PR branches would not trigger any builds at all (but with the current workflow such branches are rarely used).