add a `chain` property to the resolver result
The chain property contains the path of the route that has returned the result in the routes tree (an array of routes from the root of the routes tree to a leaf).

If the result is returned by an action which obtained it by calling `next()` the returned chain contains only the path to that action's route.