Activate all routes. (#139)
preparations for #23
---
* add a `chain` property to the resolver result

The chain property contains the path of the route that has returned the result in the routes tree (an array of routes from the root of the routes tree to a leaf).

If the result is returned by an action which obtained it by calling `next()` the returned chain contains only the path to that action's route.

* Activate all routes.

* Allow processing routes with empty paths.

* Fix merge leftowers.

* reivew fixes

 - explicitly skip the synthetic root route instead of all '' routes
 - compare routes for reference equality instead of comparing route paths
 - always initialize context.__resolutionChain

* Rework inactivate approach.

* Update action tests.

* Fix promise handling, add test.

* optimized the construction of `context.chain`

Construct the chain once after resolve is done by walking up the routes tree from the resolved route to the root, instead of updating the chain all the time during the routes tree traversal.

* address review comments

* Remove unused property.