remove build-specific Vaadin.Router import from demo snippets (#213)
* remove build-specific Vaadin.Router import from demo snippets

 - replace it with a comment showing both import and UMD options

* remove `<script type="module">` from demo snippets

the default mode in demo snippets is using the UMD bundle

* unify Vaadin.Router imports shown in demo snippets

Include two options everywhere: modules for bundlers, and UMD for browsers. Both are commented out - the users are expected to uncomment one.

* unify router outlet tagnames in demos (use divs)

Do not use the `<main>` tags to not assume anythign about the app structute, and to not confuse devs who might not know those tagnames.

Keep `<main>` if it is used alongside with `<nav>`.


Closes #67