fix: avoid URL-confusing default pathname in baseUrl (#565)
Fixes #517

When not set explicitly, the router constructor discovers `baseUrl`
using the `<base href>` value and initial document URL. However, only
the pathname from the discovered base URL was taken for the default
value of the `baseUrl` property. In case of initial pathname starting with
double slash `//`, such a value was then throwing `Invalid URL` error when
used as the URL constructor’s first argument.

This changes the default value discovery of `baseUrl`, so that it uses
an actual URL instead of only pathname to avoid throwing errors.