fix: npm install with npm 7 (#592)
npm install was failing with npm 7 unless using the --legacy-peer-deps
flag. Remove dependency to old version of eslint-config-vaadin
and stylelint-config-vaadin while adapting the configs to try to keep
the rules mostly the same for now. We might consider using the latest
version of eslint-config-vaadin and stylelint-config-vaadin again when
we migrate away from bower and HTML Imports and maybe migrate
this repo to use TypeScript.

Fixes #541.